+++ ttnn/cpp/ttnn/operations/data_movement/CMakeLists.txt
@@ -81,7 +81,6 @@ target_sources(
             reshape_view/device/device/dataflow/reader_reshape_tiled.cpp
             reshape_view/device/device/dataflow/writer_reshape_tiled.cpp
             reshape_view/device/device/rm_reshape_interleaved.cpp
-            reshape_view/device/hostdevcommon/common.hpp
             tilize/device/kernels/dataflow/reader_unary_stick_layout_split_rows_interleaved.cpp
             untilize/device/kernels/dataflow/reader_unary_start_id.cpp
             untilize/device/kernels/dataflow/writer_unary_stick_layout_split_rows_multi_core.cpp
