+++ ttnn/cpp/ttnn/operations/data_movement/reshape_view/reshape.cpp
@@ -213,13 +213,43 @@ ttnn::Tensor reshape_tiled(
             MemoryConfig{TensorMemoryLayout::INTERLEAVED, working_output_memory_config.buffer_type()};
     }
 
-    auto output_tensor_3d = ttnn::prim::reshape(
-        tensor3d,
-        requested_shape_3d,
-        requested_padded_shape_3d,
-        working_output_memory_config,
-        recreate_mapping_tensor,
-        sub_core_grid);
+    // get the variables for compressing the mapping
+    auto input_shape = tensor3d.logical_shape();
+    auto output_shape = requested_shape_3d;
+    auto tile_shape = tensor3d.tensor_spec().tile().get_tile_shape();
+    auto face_shape = tensor3d.tensor_spec().tile().get_face_shape();
+    const uint32_t num_input_pages = tt::div_up(tensor3d.physical_volume(), tile_shape[0] * tile_shape[1]);
+    const uint32_t num_output_pages = tt::div_up(requested_padded_shape_3d.volume(), tile_shape[0] * tile_shape[1]);
+
+    auto compressed_map = reshape::detail::compute_reshape_map(
+        num_input_pages, num_output_pages, tensor3d.logical_shape(), requested_shape_3d, tile_shape, face_shape);
+    // find total_num_cores
+    const uint32_t total_num_cores =
+        tensor3d.device()->compute_with_storage_grid_size().x * tensor3d.device()->compute_with_storage_grid_size().y;
+    auto estimated_num_cores = std::min(num_output_pages, total_num_cores);
+    // calculate estimated size of rt args
+    size_t rt_args_size = 4 + compressed_map.pattern_templates.size() * 4 +
+                          compressed_map.page_pattern_runs.size() * 10;  // 4 header + pattern templates + runs
+    const uint32_t max_rt_args = 341;
+    Tensor output_tensor_3d;
+    if (rt_args_size / estimated_num_cores > max_rt_args) {
+        // run untilize + reshape + tilize
+        auto untilize_tensor =
+            ttnn::to_layout(tensor3d, ttnn::ROW_MAJOR_LAYOUT, tensor3d.dtype(), tensor3d.memory_config());
+        auto reshaped_tensor = ttnn::reshape(untilize_tensor, requested_shape_3d);
+        output_tensor_3d =
+            ttnn::to_layout(reshaped_tensor, ttnn::TILE_LAYOUT, reshaped_tensor.dtype(), working_output_memory_config);
+        printf("composite path\n");
+    } else {
+        printf("direct path\n");
+        auto output_tensor_3d = ttnn::prim::reshape(
+                            tensor3d,
+                            requested_shape_3d,
+                            requested_padded_shape_3d,
+                            working_output_memory_config,
+                            recreate_mapping_tensor,
+                            sub_core_grid);
+    }
 
     if (memory_config.is_sharded()) {
         TT_FATAL(!sub_core_grid.has_value(), "Sharded reshape does not support sub core grid specification\n");
